Faster classification of sites in basins
