+++ mapbox/vs30/scripts/basin_z_values/readme.md
@@ -1,106 +1,76 @@
 # Preparation for Vs30map update
 
 This is an instruction on how to generate a file containing points *inside* known basins, and their Z-values.
-The points are from 100m x 100m grid (hard-coded in gen_basin_stats.py), and we will be using v2.07 basins.
+The points are from 100m x 100m grid (hard-coded in gen_sites_in_basin.py), and we will be using v2.07 basins.
 
-Tihs manual is based on the assumption that you are using NeSI Maui.
-It takes about 3 days to complete the steps given in this manual.
-
-
+This manual is based on the assumption that you are using NeSI Maui.
 
 ## Generation of grid points within basins
 
-(Duration: 2 days 7 hours)
+(Duration: 15~30 mins)
 
-If there are more basins added, update basins.py in Velocity_Model
-
-Also find the location of your basin files (eg. Wellington_Polygon_Wainuiomata_WGS84.txt).
-We need the root Velocity-Model directory that contains Data.
-eg. 
-If basin files are unde  /scale_wlg_persistent/filesets/home/baes/Velocity-Model/Data/Basins/,
-VMDIR will be /scale_wlg_persistent/filesets/home/baes/Velocity-Model
-(it must be an absolute path)
+Velocity_Model/Data has basins.yaml containing the list of basin models. 
+Make sure this file contains all the basins you want, and their outline files are correctly listed.
 
 ```
-sbatch --export=ALL,VMDIR=/scale_wlg_persistent/filesets/home/baes/Velocity-Model gen_basin_stats.sl
-
-```
-This will take 2 days and about 7 hours using 1340 cores on Maui. As Maui has 24 hours wall clock limit, you will need to resubmit twice after the 24 hours of run.
-Simply enter the same command, and it will resume from the checkpoint.
-
-To confirm if everything is completed, go into `outdir` directory and enter this command
-
-```
-$ cat checkpoint* |sort -n -r |tail
-121591 121591
-121591 121591
-121591 121591
-121591 121591
-121591 121591
-121591 121591
-121591 121591
-121591 121591
-118999 121591
-117999 121591
-```
-The first number is the last checkpoint index and the second is the last index that each core has to process. So the output above means the last 2 processes are still running.
-
-When everything is completed, go into `outdir` and merge the result.
-```
-cd outdir
-cat basin_stats_????.ll > ../basin_stats.ll
-cat basin_stats_????.csv > ../basin_stats.csv
-cd ..
-```
-
-The generated files look like
-
-```
-#(!524) $ head basin_stats.ll
-#168.2632171899503 -44.71216293699213 1224800_5038100
-#168.26329053597905 -44.71126577833483 1224800_5038200
-#168.26336387858157 -44.710368619441034 1224800_5038300
+- version: 2.07
+  basins:
+  - id: 26
+    name: Hawke's Bay
+    files:
+      - Basins/Napier_Hawkes_Bay/v21p7/HawkesBay1_Outline_WGS84_delim.dat
+      - Basins/Napier_Hawkes_Bay/v21p7/HawkesBay2_Outline_WGS84_delim.dat
+      - Basins/Napier_Hawkes_Bay/v21p7/HawkesBay3_Outline_WGS84_delim.dat
+      - Basins/Napier_Hawkes_Bay/v21p7/HawkesBay4_Outline_WGS84_delim.dat
 ...
 ```
-```
-#(!525) $ head basin_stats.csv
-#168.2632171899503, -44.71216293699213, 1224800, 5038100, True, WakatipuBasinOutlineWGS84.txt
-#168.26329053597905, -44.71126577833483, 1224800, 5038200, True, WakatipuBasinOutlineWGS84.txt
-#168.26336387858157, -44.710368619441034, 1224800, 5038300, True, WakatipuBasinOutlineWGS84.txt
-...
-```
+Enter the following command.
 
-To keep the job size manageable, we split them into 80k lines long files
+`python gen_siutes_in_basin.py --remove_not_in_basin`
 
-```
-mkdir -p splits
-cd splits
-split -l80000 -d -a4 ../basin_stats.ll basin_stats.ll_
-cd ../
-```
+This will generate 100m * 100m grid points, filter out offshore locations, and determines if a site is inside one of the 
+specified basins.
+The output is stored (by default) as "sites_in_basin.csv" The `--remove_not_in_basin` option removes the sites not in 
+any basins from the output file. This takes about 15 minutes on a good desktop PC.
 
-In our case, this creates 80 splits named `basin_stats.ll_0001` etc, but this number can vary. Check how many splits you have.
+If you have a station file containing station name, longitude and latitude, you can use it as an input.
+You will need to specify the index for these columns as below. 
+
+`python gen_siutes_in_basin.py --stat_file station_file.csv --stat_name_col 0 --lat_col 1 --lon_col 2 --sep , --skiprows 1`
+
 
 ## Computing Z values for given locations
 
-(Duration: 6 hours)
+(Duration: 3~4 hours)
 
 We will be computing Z values using NZVM. This requires quite a large memory (96Gb peak), and it is recommended to run on Maui using the supplied SLURM script.
-From above, we have 80 split files to process - if you have a different number, update the SLURM script.
 
-In particular, you will need to update "79" below.
+Let's split the file into smaller files, with 80000 lines each.
+
+```angular2html
+(ringo310) seb56@c015kr:~/mapping/mapbox/vs30/scripts/basin_z_values$ mkdir splits
+(ringo310) seb56@c015kr:~/mapping/mapbox/vs30/scripts/basin_z_values$ cd splits/
+(ringo310) seb56@c015kr:~/mapping/mapbox/vs30/scripts/basin_z_values/splits$ split -l80000 -d -a4 ../sites_in_basin_whole_nz.csv sites_in_basin_whole_nz.csv_
+(ringo310) seb56@c015kr:~/mapping/mapbox/vs30/scripts/basin_z_values/splits$ ls -ltr
+total 407008
+-rw-rw-r-- 1 seb56 seb56 7911070 Sep 29 15:46 sites_in_basin_whole_nz.csv_0000
+-rw-rw-r-- 1 seb56 seb56 8045492 Sep 29 15:46 sites_in_basin_whole_nz.csv_0001
+-rw-rw-r-- 1 seb56 seb56 8102851 Sep 29 15:46 sites_in_basin_whole_nz.csv_0002
+-rw-rw-r-- 1 seb56 seb56 8119448 Sep 29 15:46 sites_in_basin_whole_nz.csv_0003
+...
+-rw-rw-r-- 1 seb56 seb56 9007784 Sep 29 15:46 sites_in_basin_whole_nz.csv_0044
+-rw-rw-r-- 1 seb56 seb56 9001716 Sep 29 15:46 sites_in_basin_whole_nz.csv_0045
+-rw-rw-r-- 1 seb56 seb56 9004491 Sep 29 15:46 sites_in_basin_whole_nz.csv_0046
+-rw-rw-r-- 1 seb56 seb56 5924069 Sep 29 15:46 sites_in_basin_whole_nz.csv_0047
+
 ```
-#SBATCH --array=0-79%4 # 80 array jobs with 4 jobs running at a time
+From above, we have 48 split files to process - if you have a different number, update the SLURM script.
+
+In particular, you will need to enter as "47" below.
+```
+#SBATCH --array=0-47%4 # 48 array jobs with 4 jobs running at a time
 
 ```
-
-The following variables will be automatically set.
-```
-$SLURM_ARRAY_TASK_ID  0-79
-$SLURM_ARRAY_TASK_MIN 0
-$SLURM_ARRAY_TASK_MAX 79
-````
-
 
 Find your VMDIR as above and replace it in the below example
 
@@ -114,14 +84,14 @@ This will keep 4 array jobs running at a time, each job takes less than 30 mins.
 When everything is completed, merge all the output files into a single file
 ```
 cd z_values
-cat basin_stats.z_???? > ../basin_stats.z
+cat sites_in_basin_whole_nz.z_???? > ../sites_in_basin_whole_nz.z
 cd ..
 ```
 
 Finally, run the command below.
 
 ```
-python combine_basin_stats_z.py basin_stats.csv basin_stats.z
+python combine_basin_stats_z.py sites_in_basin_whole_nz.csv sites_in_basin_whole_nz.z
 ```
 
 This must be quick. It will create `basin_stats_z.csv` file.
